just use produtPage fixture for all product page tests
We were calling both the `productPage` fixture and the `page` fixture in
some tests. This made the test fixture use signurate a bit bloated.
The only reason we were doing this was to use the `page` methods that
are not direclty available on the `productPage` fixture. However,
the `productPage` fixture has a `page` property that is the same
instance as the `page` fixture. So we can just use this property to
call any of the `page` methods.

Additionally, there were other tests that were using the `page` fixture
for product page. Let's use the `productPage` fixture for all of them.
This will be more helpful when we update the `productPage` fixture to
use the custom server port for the base URL.