change the base url for the product page fixture
When we use the custom next js server we need to use a different port
for the `goto` method. Previously we were importing the port fixture
and using the `goto` method from the `page` fixture to navigate to
the explicit path on the new port. If we add a base url property to
the productPage fixture, then we can modify it when we call the
custom server fixture.

As such we can remove the need of having to import the `port` fixture
directly. However, we will still keep the `port` fixture as there maybe
tests that need access to the custom server without needing to use
the productPage fixture.