add data-testid for out of stock test
We are targeting elements by their text, which is not a good practice.
By adding a data-testid attribute, we can specify the element we want to
target as well as test if the element is visible. We do not necessarily
need to test the text inside the element.

This will make the test more resilient to changes in the text.