add data-testid to cross sell section price
We want to validate the price of the cross sell section, but by using
'Total price' as part of the selector, we might end up selecting a
different price element. Let's add a data-testid to the total price
element of the cross sell section, so we can target it directly.