add data-testid for product info section
For testing certain product information, we target elements by their
text content but scout the entire page for these elements.
This is not ideal because we could end up targeting the wrong element.
Instead let's add some data-testids to specify certain elements.
In addition, let's narrow down the search of these elements by
targeting the product info section. This will be helpful when we it is
not possible to add data-testids to the elements but do not want to
search the entire page for the element.