create a response type for each request handler
This will reduce the amount of type assertions to distinguish between
a response for a GraphQL request and a response for a REST request.

A GraphQL response is only valid if the the body is a JSON object, while
a REST response can have any body type as well as headers. However, if
a body is provided for the REST response, then a response type for the
transformation must also be provided. As such, by creating a new type
to define this body and response type co-dependency, we can let
Typescript do the heavy lifting for us and reduce the amount of type
assertions.

Co-authored-by: Andrew Gilbert <andyg0808@users.noreply.github.com>