separate handler options for each handler type
Following a similar pattern as the previous commits, let's separate out
the handler options for each handler type. This will reduce the
amount of type checking we need to do in the handler creation code. In
addition, let's declare the custom resolver as a separate type so
we can express to the user that passing a custom resolver will ignore
the `once` and `passthrough` options.

Co-authored-by: Andrew Gilbert <andyg0808@users.noreply.github.com>