explicitly use node 16 for runners
The default node version for the runners was changed to 18, which
is not compatible with our current setup. Unfortunately, this change
was not properly announced, which is why we did not notice it. There is
only a GH issue about it at the moment.
https://github.com/actions/runner-images/issues/7002

In order to fix this, we need to explicitly set the node version to 16
for all our workflows that use node/pnpm.