fix(cross_sell test): narrow down selector to avoid grabbing unwanted text
We were grabbing the text of the entire cart drawer body which included
the text of the upsell product component. This became a problem when the
upsell product was one of the products of the cross sells. As such,
when we were validating that the cross sell product title was not in
the cart drawer body we would find it as part of the upsell product and
fail the test.

By narrowing down the selector to only grab the `li` elements we will
avoid grabbing the text of the upsell product component. In doing so,
we need to update the `cartDrawerText` variable name to `cartDrawerItems`
to better reflect what it is. Furthermore, we need to update the
assertion to  `toContainText` instead of `toContain` as the latter
only works with a single string and the former can work with
an array of strings.