scroll cross sell section into view
When looking at the trace result for the cross sell test, it seems that
the cross sell section was not visible on the test screenshots. Although
the test was still able to interact with the cross sell section, we could
view these interactions on the screenshots. Therefore, let's scroll
the cross sell section into view before interacting with it.

This should help make it easier to debug the cross sell test in the
future.