Specify path for playwright ts
Playwright tries to read all of the other React style file in the
package which is doesnt know how to read so there was a complilation
error. So, we have to specify specificially that we only want to import
the typescript code so we dont import react code. This is a temporary
fix.

Co-authored-by:  Angel <angel@ifixit.com>