Memory Cache: don't clear on vercel_no_cache
We are using Alertra to monitor uptime and we add a _vercel_no_cache=1
query param to ensure we actually run app code. We don't want this to
constantly clear the in-memory cache because we have alerta checking
twice a minute.

The in-memory cache only stores a few global settings and it expires
every hour.