deflake parts page search test
This test is flaky as we wait for the algolia query to finish but this
event does not mean that the page has fully been updated. As a result,
we sometimes fail since we have not updated the url after the query has
finished.

To fix this, we will instead wait for the url to be updated with a query
parameter after will enter the search text but before the assertion
is made.

Co-authored-by: Federico Badini <federicobadini@users.noreply.github.com>