modify filters test to use new algolia results structure
The Algolia results structure seems to have to changed recently and
therefore our previous assumption on the results structure is no longer
valid and thus cause the test to fail.

The Algolia results now include hits from different filter parameters.
We need to search for the filter params that include the facet option
we selected. This will indicate that the hits are associated with
the filter facet option we selected.

Additionally, there are some gaurd clauses to ensure that if our current
assumption about the Algolia results are no longer valid,
we will throw an error to indicate that the test might need to be updated.

Co-authored-by: Mason McElvain <masonmcelvain@users.noreply.github.com>