Troubleshooting: Add author name as description text on avatar
This avoids accessibility problems on the avatar. It's redundant with
the text right next to it, but I think that's not the biggest problem,
and the Chakra component doesn't seem to allow me to pass through the
`aria-labelled-by` attribute to the underlying image (rather than the
wrapping element), so there doesn't seem to be a simple
alternative. We could probably make something work (involving a custom
component and and `as` prop), but I just don't think the slight
incorrectness of having the author's name in text and on the avatar is
the end of the world.