Troubleshooting: Clear around Lite YouTube videos
Now that we're using custom elements for these, we don't have the
`.videoBox_center` element to target. We mostly just want to clear
these for now, because it keeps the text below here on the demo page
from wrapping around the floated images above (which then makes it
look like the images are rendering differently).