Add stress-test workflow initial setup
In the initial setup, we will check the diff between the base ref sha
and the head sha to check if any test files were changed for playwright
and jest tests. If they were changed, we will set outputs for each of them
and trigger the next jobs to run the tests.

Also copied over the skip stress-test logic from iFixit/ifixit.