Troubleshooting: Add scroll to nav tabs
This enables them to scroll when there's not space for them to render
normally. In order to avoid needing JS to implement this, I decided to
use just the padding width as the gradient, so we dissolve the tabs
into the padding around them. This then allows the `::before` and
`::after` pseudoelements to hide because they blend into the padding,
which scrolls away behind them when the user starts to scroll the tab
bar. They also disappear on screens large enough that there shouldn't
be a need for the scrolling.