clean up running playwright tests section
This organizes the information for running playwright tests a bit better.
I am also going to create a new file to document how we mock API requests
and more information about how to debug playwright tests. I chose to
create a new file because the README was getting a bit long.