create new doc for elaborating on testing with playwright
Instead of creating bloat in the README, I've created a new doc for
elaborating on testing with Playwright. This will be a good place to
add more information about testing with Playwright in the context of
our application.

For now, I've elaborated on the mocking features we've implemented and
how and when to use them.