extract mocked layout props to a separate file
Following suite as previous commits, this commit extracts the mocked
layout props to a separate file. By doing this we can add additional
mocks or utility functions within the file while retaining the context of the mocks. In addition, since this mock is for a hook, we
can add the `tsx` extension to the file name to make it more clear
that this file either uses or imports a React component. Which caused issues as listed in https://github.com/iFixit/react-commerce/issues/1400.