Breadcrumbs: Fix wrapping bug
Using `min-content` here meant that the excess breadcrumbs don't get
wrapped into the hidden overflow and disappear. Instead, let's use a
proportion of `em` that seems to result in the tails of the letters
being visible. I'm figuring that, because em are font-size relative,
this should be resiliant against font size changes.