create a new fixture to override the trace output path
There is no simple way to format the trace file output path like the
`snapshotPathTemplate` option in Playwright. The second most simple way is
to override the `testInfo.outputDir` property. This is a bit hacky but
the alternative is to hook into the `tracer` api in a custom test runner.

The actual override will be done in the `test-fixtures.ts` file, but this
fixture will generate the template path for the override. The template
will be as follows:
`[subtest-directory]/test-file-name/test-name/device-name/` ->
`product/add_to_cart/back-to-shopping-button-works/desktop-chrome/`

This nested directory structure will allow us to easily find the trace
files for a given test as the original trace files were underneath a
single directory where the directory name was a flattened version of the
new directory structure.