apply the new trace fixture override to all tests
Now we can use the new trace fixture override method to set the `outputDir`
for `testInfo`. By also setting the `auto` option to `true`, we do not
need to include the `testInfo` fixture in the test signature to make
use of the override, it will be automatically injected to every test.