Admins: add idea of a server-side adminMessage
This allows the server to show a message to just Admins. The string will
be in the props reagrdless, so we can't send anything in secret.

This was mainly added to deal with syntax errors in the algolia filters.
But, it can be extended to del with any server-side errors that we'd
like to expose to admins.