fix: add test suite to trace output path
We added this trace fixture in https://github.com/iFixit/react-commerce/pull/1563,
but while addressing https://github.com/iFixit/react-commerce/issues/1569
I noticed an issue with the trace output path. There are some test cases
that have similar names, but are under different test suites. This
causes the trace outputs to be overwritten.

To fix this, we can take full advantage of the `titlePath` property to
create the trace output path. The `titlePath` property is a convenient
array of strings that represents the relative path to the test case from
root test directory. This includes the file name, the test suite tree,
and the test case name.

In addition, we can refactor the `formatTestNameForDirectory` function
for more general usage. As such it will now be called `formatName`.