refactor: rename test names to be more descriptive and accurate
This came up in a comment on a PR -
https://github.com/iFixit/react-commerce/pull/1563#issuecomment-1505765586

There are a few tests that have names that some special characters in them.
By removing some of these special characters from the test names, it
provided an opportunity to make the test names more descriptive and accurate.

I used ChatGPT-4 to suggest the new names given the playwright code. I
then manually reviewed the suggestions and made the final changes.

The test names should now be general enough but also specific enough to
describe what the test is doing without using any special characters.