update comment finding logic for nextjs bundle analysis
The nextjs-bundle-analysis package was updated, resulting in a change
to the comment text that the existing search logic was trying to find.
The relevant change can be found in the following commit:
hashicorp/nextjs-bundle-analysis@067cb5d

Let's update the search logic to match a comment on a viewable string
and by the github bot user.