use specific version for nextjs-bundle-analysis package
By using npx we will always download the latest version of the package.
As a result, we will end up using a version with breaking changes to
our CI workflow. Let's use a specific version to avoid this issue and
since we fixed the workflow with the latest version, we can use it - v0.5.0.