Remove variantSku conditonal
We don't need this anymore since previously, variantSku corresponded
to the selectedVariant on the page which could've been null.

However, now we use the sku for the product that the review is
about instead of the variant Sku so we don't tneef the conditional.
Moreover, we got rid of the prop entirely, so let's get rid of this
check also.