Prerendered: Use `fontSize` to control the font size of the text
This uses a default `fontSize` value on the `Box` instead of providing
a `fontSize` for the `p`, which means that the `p` will inherit the
`Box`'s font size. As a result, the `chakra` `fontSize` prop on the
`Prerendered` component will also affect the font size for the text
within it. This allows us to configure the base size of the text in
the component, rather than having it constantly set to 16px.