fix(product cache): drop type checking
The Content-Length header is not necessary for localhost, but is strictly necessary for Vercel deploym
ents. Otherwise, the request.body does not end up containing the right data.

Drops the type checking of the request body in the handler. The body will always be a json object due
to the headers provided in the request.