refactor(compatibility): update linking
Updates product overview (aka sidebar) compatibility links to point to the compatibility section of the page, instead of the device page.

Updates the compatibility section cards to open/close the dropdown when clicked, instead of linking to the device page. The image and title still link to the device page. Clicking the model numbers will have no effect, so that users can copy/paste them without toggling the dropdown.

Replaces the disclosure with a details element to enable find-in-page in chromium browsers.