fix(compatibility): conditionally include links
Only renders anchor tags in a compatible device card if a device url is provided.

In the product overview (aka sidebar), we want the whole card to link to #compatibility. In the compatibility section, we want the image and the title to link to a device page.

The fragments in CompatibilityWrapper are used to appease typescript. See more here: https://github.com/DefinitelyTyped/DefinitelyTyped/issues/18051