ifixit-api-client: Include `Authorization` header on server-side requests
This allows access to our dev server from the server-side. We're not
using a cookie here because I feel like an ordinary header is a
cleaner way to implement this (cookies are a multi-value string for
one header and might require merging, while a single-purpose header
can just be used).