Playwright: Generate cookie in config
We were having problems because Playwright's project filtering stops
running the dependencies of the project which is being filtered to,
meaning that the `setup` project never gets run.

Turns out that we don't need to provide a JSON file of auth
information. The API accepts hard-coding the cookie in the
config. This means we don't need the `setup` project at all and we can
simply move the same cookie generating code into the config and use
the environment variables from there.