Change the text from reviews to ratings
As specified in the issue, when the page displays a number it is followed
by reviews but the number also takes into account the ratings given to
the product. Therefore, it makes sense to change this text accordingly.

From my understanding, the `totalReviewsCount` is the value
we get from `reviewsData.count` which probably counts the number
of ratings instead of written reviews.