delay mocked responses for 2 seconds
There was an issue with the msw request handlers for `strapi` requests.
https://github.com/ifixit/react-commerce/issues/1627

After digging into the issue, it seems to be an issue with `node-fetch`
or `strapi` itself. The trigger for it was 3 requests being made directly
to `strapi`: `strapi.getGlobalSettings`, `strapi.getStoreList`, and
`strapi.findStore`. These requests were not matched by any of the msw
request handlers, so they were passed through to the real `strapi` server.

Then a `strapi.getProductList` request was made, which was matched by the
msw request handler, so it was intercepted and mocked. However, this
request was immediately returned while the previous 3 requests were still
pending. This caused the `FetchError: socket hang up` error to be thrown
by `node-fetch` as the queue of pending requests were closed since
the lastest request was already returned. In order to fix this, we need
to delay the mocked responses to ensure that the previous requests have
been returned before the next request is mocked. This also mimics
the behavior of real requests, which are not returned immediately.