Difficulty badges: Match colors and icons from Guide pages
This was annoyingly copied from Guide Products, with a modification to use difficulty names instead of numbers.
That also means that this breaks once we start using translations (it should default to the moderate difficulty style).
Also note that the badges are using Font Awesome `pro-solid-svg-icons` here instead of `pro-regular-svg-icons` like on Guide pages (the mockup uses solid icons).