Algolia: count facet results after doing distinct
We'll now be using the "attributeForDistinct" feature. This means that
for some products there will be more than one algolia document that
represents it. We'll be turning on this attributeForDistinct feature which
will ensure we only get one result per product in the result set.
However, the result counts in the facets will show the number of
documents instead of the number of products until we set this option.

See: https://www.algolia.com/doc/api-reference/api-parameters/facetingAfterDistinct/