Algolia getServerState(): optimize what's needed
getServerState() does a pre-render phase and then extracts the result
set and facets from the internal state so they can be included /
captured in the props.

Because of this, we don't need any of the rest of the page, just the
minimal UI needed to fire off the correct query. Here we add a
"minimalMode" that tries to use the same hooks and filtering, but render
as few UI components as possible.

In testing, this dropped about 100ms of component time from *everY8
request while providing the same results.