useHits(): get hits info from SearchDetailsContext
useHits() is expensive (> 1ms per call) and we called it in quite a few
places. Calling it once instead saves ~10ms or so.

Note this savings happens twice on the server (in getServerState() and
in the next.js render-to-string phase, then it happens client side after
evey change in the search query.