fix(codegen): pregenerate schema file
Uses codegen's programmatic api to save the strapi graphql schema to a file before exporting the rest of the configuration.

Motivation:

The validation schema generation depends on saving the Strapi graphql schema to a file first. However, codegen seems to perform all generations in parallel, so it fails to generate the schema file before the validation file.