useHits(): prefer usePagination() where possible
useHits() is an expensive call on the server (> 1ms per call) and we
called it in quite a few places (some are in loops). Calling
usePagination() is < 0.1ms and provides us the thing we need most often
(hit count). This saves us ~10ms or so.

I've noticed that useHits() is not slow on the client strangely enough.

Note this savings happens twice on the server (in getServerState() and
in the next.js render-to-string phase.