Upload test result file generated from metrics reporter
After tests finish running, the metrics reporter will generate
playwright-result-${this.commit} file in the frontend directory
for each project.

Then we rename the files with their matrix project name to ensure that
it doesn't overwrite each other since initially they will have the same
name. The resulting files will be named as `playwright-result-$BUILD_COMMIT-$matrix_name`
in the frontend folder of the react-commerce repo.

Then, they will be uploaded and downloaded to playwright-results/ directory in the second job.
For example:
- playwright-results/
    - playwright-result-$BUILD_COMMIT-chrome-desktop
    - playwright-result-$BUILD_COMMIT-firefox-desktop...