Record playwright test metrics
Added a new job that is set to run after playwright-run jobs finish on
success or failure.

If all of them are cancelled, or skipped the artifacts will not be uploaded
so don't run on cancelled() or skipped().

This job checks out iFixit/ifixit and downloads the artifacts in
playwright-results/ directory. Then it combines all of the results
from different matrices in playwright-results folder.

Then that file is fed into Tests/Exec/report-test-run as stdin.