Troubleshooting: Redirect to canonical URL when needed
This was tricky, because I didn't want to redirect to the canonical
hostname as reported by the server, in order to keep preview versions
of the NextJS site working correctly. There's also no protocol or
hostname in the `resolvedUrl` value, so it's trickier to know what the
correct values for those would be (particularly the protocol, since
the hostname is at least available as a header). I've decided to make
the assumption that `localhost` is `http` and everything else is
`https`.

Ref: https://nodejs.org/api/http.html#messageurl
     The technique they show for constructing a URL from the URL Node
     makes available demonstrates that there's not an easy way to get
     the protocol, and that the `host` header is the best we get for
     the correct hostname.