Troubleshooting: Redirect to relative URL
The `Location` header supports relative URLs; let's use those to allow the
browser to handle the protocol/host matching for the redirect.

This code seems to have been causing weird redirect problems in prod;
I'm not sure what's going wrong, but simplifying it seems to be a good
first thing to try.