Troubleshooting: Redirect on canonical path mismatch
We were comparing the `resolvedUrl`, which we presume would include the query
string, with the pathname of the `canonicalUrl` which doesn't.

Now we extract just the pathname from the `resolvedUrl` and compare that.

This also ensures that we include the query string from the current URL
when we construct the target URL for the redirect, which I realized we
were missing.