Unskip msw tests that use createGraphQLHandler
Since there is no way to mock the server-side api calls on next 13,
we are going to using some hard coded products.

Most of these products are stocked by a cron job on iFixit/ifixit side
in StockPlaywrightTestProducts class, so it's a bit safe but not a permanent
solution, since a data change on Akeneo could cause some of the tests to fail.

But for the moment, this let's us going and have some test coverage until we
merge this repo to the monorepo.