Troubleshooting: Extract function to fetch data from server
We're going to extend this to support multiple possible ways to look
up the data. Having this function allows us to fall back through all
the options. I think it's a cleaner way to handle the task anyhow, as
it avoids duplication and mutation to handle the try-catch needed to
handle fetch errors.