Remove nvm use from script
We don't want to force the user to have nvm as the version-er for node
so lets remove it from the script and make the user call it outside of
the script.

We need access to the bash resources for nvm use, so this now lets us
use ./ to run the script in a separate environment rather than using
source to run.