update vulcan playwright tests
We made a change in the iFixit/ifixit repo which replaced spaces in
urls with '+' characters (https://github.com/iFixit/ifixit/pull/48898).

As a result, the playwright tests in the `react-commerce` repo have
started to fail because we were matching with `%20` instead of `+`.

This updates the tests with the new url format and also dropped an
unnecessary `await` in the `loading.spec.ts` test.