Hreflangs: Get rendering
The hreflangs component was working as expected. The correct links were rendered,
but when viewing the page no links were actually rendered. It ends up we
are running into an edge case in next/head: https://github.com/vercel/next.js/issues/8384

This commit adds a workaround for this issue which was suggested from the
aforementioned issue.

The fix is to invert the usage such that <Head> is used in your component.
Muliple uses of <Head> will be merged together.