Cart: copy cart drawer implementation
We're going to re-use the useCart() hooks for state and mutation, but
the UI will likely be different. Probably not all that different though,
so let's try starting with a copy of the cart drawer code.

Further commits will remove un-needed bits and fix the relative paths
and such. This commit is a direct copy/pasta.