Refactor: Cache control easier configuration
This commit refactors the cache control helpers to make it easier to
configure the cache control headers.

I had noticed that disabling the cache control headers was implemented in
a way that was not very intuitive. In fact, the disabling code exists in
another file / middleware.

This commit makes it easy for us to move the disabling code to the
this file.