iFixit API: Force no caching
OMG Next overrides fetch and defaults to caching:
https://nextjs.org/docs/app/api-reference/functions/fetch

Worst of all, they drop many of the configuration values.

As far as Danny and I could tell we would want the mdn value of "default"
but Next does not support this value.

We basically get "on or off" and "For how long?" as a global cache even between
deployments on vercel.

Lastly we may need to discuss overriding fetch globally in next to avoid this
problem in other places.