noindex all responses
@jarstelfox and I discovered this while testing something else and came
up with this solution.

The short of it is, since vercel caches responses based on url only
(seemingly), we can't have it changing the response based on a header.
Currently, we have code here to add a "X-Robots-Tag: noindex"
header if there are Cloudfront headers on a request. This works some of
the time but we end up caching (and serving) responses generated for the
same url, but different sets of headers.

Effectively, the requests from CF and the the requests directly to
https://react-commerce-prod.vercel.app share an http cache in vercels
CDN. Because of this, we need to not change our vercel responses based
on headers as they aren't part of the cache key.

The most straightforward way to disallow indexing on all requests except
for those coming through CF is to always add the no-index directive and
expicitly have CF remove it from all responses it deals with.