Troubleshooting: Add a scroll percentage indicator to the page
This is a simple component that shows the scroll percentage of the page.
It takes an optional `scrollContainerRef` prop that will limit the indicator
to the scrollable area of the container.

If not provided, it will default to the height of the window.

When using the ref, we will also stay at 0 until we have scrolled past the
top of the container.

This is useful since we are rendering this below the header, and we don't
want to show the scroll percentage of the header. Only the scroll percentage
of the content.