Scroll Indicator: Add some options
I was messing around and wasn't sure exactly the behavior we might want.
So I added some options to the scroll indicator. We can choose to always show
the indicator (on page load it takes up space and looks odd)
or we can choose to only show it when the user scrolls.

Similarly, we can choose to show the indicator after the user moves past the content
or we can hide it after the user moves past the content.