TOC: Create a tocContext
This does not handling of rendering, but instead just handles the main logic
of the TOC.

You push on a ref object and it's title in the TOC. The context will
automatically watch the scrolling and update to the item whose ref is
currently in view.

Currently in view is defined as the ref is in the view port and the largest
portion of the ref is in the view port.

This means a tiny 50px div in the center of the viewport will be considered
in view, but a 1000px div that is 1px off the top of the viewport will not
be considered in view.

This is a good compromise between the two, but may need to be tweaked.

The context also provides a way to get the toc items (name / ref pair)
and will automatically sort them by the vertical position of the ref.

If the vertical position matches then it's left to right.