TOC active: Use intersectionobserver to detect which section is active
Right now we are allowing more than one. i.e. anything that is in the viewport by a pixel
is considered active.

This commit does a huge chunk of refactoring under the hood and is not
the final destination. It is a step in the right direction.

As of this commit the TOC highlights anything in view.