TOC Context: Include scrolling behavior
The way we determine if a TOC item is active is based on a custom
algorithm that calculates the percent traveled.

If we don't expose a way to scroll an item it's unclear to the implementing
developer exactly how the context selection for active works.

Now the scroll to logic is exposed and does not require the dev to read the logic.

As part of this commit I also added an optional buffer to the scroll to logic.
This is required because already have an element which takes up no space,
but then does after some amount of scrolling: The scroll indicator.

I will look to see if I can remove it from the css flow, but for now this works great.