TOC: Start to match figma
Some notes here:

This code:

```
marginLeft={-4}
listItemProps={{
   paddingLeft: 4,
}}
```

Is to break out fo the container such that we can have the background color
extend to the left of the list item.

The different exports useScrollPercentHeight allow TS to infer the type of the props.

I am working hard to get TOC / TOC Item to be a generic component that can be used
in other places. As such, I am trying to allow props to be passed in to control
the styling of the component at each call site, but have the default styling
make sense.