Server side: Render titles
It seems setState / useEffect is not called server side in nextjs. So we
need to render the titles server side.

I am thinking I will work to undo some of the more dynamic changes of this
TOC Context and instead make it such that the TOC renders server side and
then the client side code can take over.

Coming soon in near commits:

The current thought is to change pushItem to instead always push a new title
or throw an error if the title is already in the TOC.

Then we can add a new method which will get the TOCRecord for a given title.
This will be used to get the element ref for the title.

With the two above changes we can render the TOC server side and then have
the client side inject the refs.

Similarly we can have the client side actually be able to add to the TOC on the fly.
Although we will not need to use this feature right now, I am going to keep
it on the context for now.

Effectively, pushItem will be dead code, but it also may be useful in the future.