TOCRecord: Remove visibility state
This commit fixes a bug where were had two setstate calls
one for visible and another for active

They mostly ran in the same frame, but not always.

The way the mutation observer works is that it will only update when an el
is in or out of the view port, but not every scroll event.

This was trying two events together, which is why we had a bug where the
active state was not being set.

After some debugging, I found that we didn't need the visible state,
so we can just drop it.

Thus, only one setstate is called per frame and
we have data consistency.