Features: Implement a basic UI / client and server-side handling
We now can implement feature flags.

Due to how next-js works I used cookies.

The server has no access to a storage mechanism.

We could totally force you to use a query param, but
then it defeats the point of having a nice UI to toggle features.

I spent a lot of time ensuring the server side can start with no cookie
and just a query param and it will set the cookie in the response and the server
will have the correct state before the first render.

This means feature flags are effectively on or off with no
concern for serverside or not.

As of this commit nothing is using the new toc flag, but that will happen next