/Store pages: Drop catch-all route
For potential vercel proxying, we won't be able to use a catch-all
route. Since we've only eveyr used /Store, let's limit all routes to
`/Store/*` (and `/Store`).

Note: This means that "pages" in strapi will only be matched if they
start with `/Store`. No changes are needed for the existing /Store page.

This way, we can test a fallback that will proxy to ifixit.com on
unmatched pages.