App Router: Add a hello world page
This is the first use of the new app router.
It is a simple hello world page that uses the new app router to render a page with a dynamic path.

This new page will 500 right now, as we are missing some requirements for the app router.

The main requirement is a layout.tsx which is the root layout:
https://nextjs.org/docs/app/building-your-application/routing/pages-and-layouts#root-layout-required

Another requirement is that we update some hooks to a newer syntax.
These changes will come in the following commits.

One big note here is the folder layout here has (defaultLayout) in the path.
This is so that we can have a shared layout for many pages.

These are known as route groups:
https://nextjs.org/docs/app/building-your-application/routing/route-groups
The folder path does not affect your url path but allows you to organize your code better.

It also allows you to share layout/page/template files if required.