Root Layout: Add it!
I should note this is a copy from Federico's work on
https://github.com/iFixit/react-commerce/pull/1932/commits/48cd99b8bbba985ba33d2979869cc147de0fbe6a#diff-ccddbfb0d0537ee9a04df26462300003e5c30b9f0a212393cdcce2d30d2bd931

This root layout is required for the app router to work:
https://nextjs.org/docs/app/building-your-application/routing/pages-and-layouts#root-layout-required

The code inside it will be used for *every single page served by the App Router* on the site.

The current code is just polyfills and analytics. These make perfect sense
on all pages right now.