defaultLayout Pages: Wrap in iFixit header/footer
This is the main reason why I used the Route Grouping:
https://nextjs.org/docs/app/building-your-application/routing/route-groups

We can add a layout which is the header/footer and it will be used
as a layout for all pages in that group.

A couple of things to note:

1. I dropped the <ServerSidePropsProvider>

I did some digging and I believe this provider is not required for all routes.
In the few places it is included (in the old pages router) we could still add
the provider as the first component of a page and all would be fine.

2. The layout automatically fetches the layout props from the <AppProviders> context

The means the layout can be used in any page and it will automatically
get the layout props and the page can just be concerned with rendering
the non-layout components.

2.1. The layout props are not passed to the page

See: https://nextjs.org/docs/app/building-your-application/routing/pages-and-layouts#layouts

> Passing data between a parent layout and its children is not possible.
> However, you can fetch the same data in a route more than once, and
> React will automatically dedupe the requests without affecting performance.
> Layouts do not have access to the current route segment(s).
> To access route segments, you can use
> useSelectedLayoutSegment or useSelectedLayoutSegments in a Client Component.

3. A note on data-fetching

See above:
> However, you can fetch the same data in a route more than once, and
> React will automatically dedupe the requests without affecting performance.

Next takes about this at length here:
https://nextjs.org/docs/app/building-your-application/data-fetching/fetching-caching-and-revalidating#caching-data

At a high level, Next will automatically dedupe requests for the same data.

They do this at the level of the fetch API, so if you have two fetches
to the same URL and exact same options then it will only make one request.

This is true even if the fetches are in different components.

4. This may make more sense as a "Template.tsx" than a "Layout.tsx" but I'm not sure.

See: https://nextjs.org/docs/app/building-your-application/routing/pages-and-layouts#templates

> Templates are similar to layouts in that they wrap each child layout or page.
> Unlike layouts that persist across routes and maintain state,
> templates create a new instance for each of their children on navigation.
>
> This means that when a user navigates between routes that share a template,
> a new instance of the component is mounted, DOM elements are recreated,
> state is not preserved, and effects are re-synchronized.

My understanding is that a Template is a Layout that is always rerun.

If we do have some non-shared state (notifications, user logged in) and
if those are done via SSR then we may want to use a Template instead of a Layout?

That's my guess anyway.